requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
google-cloud-dialogflow==2.25.0
vk-api==11.9.9
